Move local variable naming to Codegen
This is necessary because we will need to integrate this with codegen
further (by including the sub-expression registers into the local
variable array). This means Expr has to expose the representation of
global_var_id and local_var_id as int - I previously said I was
reluctant to do this, and while I could have added functions in Expr to
convert the IDs to ints while preserving their abstraction, I didn't
think it was a big enough issue to justify the added complexity. It's
possible that it might be a good idea to do this to string_id and
proc_id as well eventually, if only for consistency.