several improvements to codegen of (call)
1. The function is no longer evaluated after its arguments. As the
now-deleted comment noted, this was technically permitted by the
standard but not what users would expect.
2. An expression like (call (call f 1 2 3)) wouldn't compile correctly
because the codegen wouldn't realize it had to allocate additional
registers to hold the inner arguments until it was too late. This was
basically a result of me forgetting that write_* functions shouldn't be
used in writer functions - I could probably check for this