fix UB from unsequenced local array initialization
Each function (and the top level) has number of `Obj` variables which
store, among other things, the values of local variables, temporaries
needed during evaluation, and the function's parameters.  Most variables
are initialized to NIL, with the exception of function parameters, which
are initialized with the macro `UNSAFE_NEXT_ARG`, which expands to
`call_args[arg_i++]`. In other words, it reads the next argument from a
global array and then increments a local counter.

Originally these local variables were declared something like this:

    Obj r0=NIL, r1=UNSAFE_NEXT_ARG, r2=UNSAFE_NEXT_ARG, ...;

C guarantees that there is a sequence point after each variable is
initialized, so this was well-defined.

However, the compiler was later changed to initialize objects like this:

    Obj r[...]={NIL, UNSAFE_NEXT_ARG, UNSAFE_NEXT_ARG, ...};

This was needed because the garbage collector needed to treat all of
these values as roots, and this allowed it to register them more easily
than pushing them all separately.

It appears that the C standard does not guarantee that a sequence point
occurs between elements of an array initializer.  Therefore that style
was technically undefined behavior. This commit fixes this by
initializing locals like this:

    Obj r[...];
    r[0] = NIL;
    r[1] = UNSAFE_NEXT_ARG;
    r[2] = UNSAFE_NEXT_ARG;
    ...