move builtin operator definitions into new file
This way we can gather all the relevant information about each builtin
into one place and avoid having to add a new ADT variant in three
different places. It will also be relevant once we start removing the
distinction between built-in and user-defined operators (such as having
to call the latter using `(call)`).

One of the knock-on effects of this change is that every operator has to
have all of its arguments evaluated before it starts doing anything.
Previously `(+ a b c)` would end up getting compiled to something like:

    out = a
    out = add(out, b)
    out = add(out, c)

And now it gets compiled to something like:

    temp1 = a
    temp2 = b
    temp3 = c
    out = temp1
    out = add(out, temp2)
    out = add(out, temp3)

This means we have to allocate more registers sometimes than we
otherwise would, but is technically more semantically correct, because
if `a + b` overflows, it shouldn't error and abort the program until `c`
is finished evaluating.

Of course, we could specifically do that by setting a flag to check for
overflow and waiting to error at the end, but that's not worth the extra
complexity at all.

Here's another thing this fixes: an expression like `(call f a b c)`
would previously compile to something like:

    temp1 = f
    init_args(3)
    temp2 = a
    arg(temp2)
    temp2 = b
    arg(temp2)
    temp2 = c
    arg(temp2)
    call(temp1)

And now it compiles to something like this:

    temp1 = f
    temp2 = a
    temp3 = b
    temp4 = c
    init_args(3)
    arg(temp2)
    arg(temp3)
    arg(temp4)
    call(temp1)

This is more correct, since each argument in the expression could
theoretically be another `(call)` expression which would previously
overwrite the old values and probably cause all sorts of UB.

I realize, as I write this, a second subtle way in which the old version
was wrong: the GC hasn't been treating the argument buffer contents as
roots. That's fine now since there's no way an cycle can happen in
between `init_args` and `call`, but previously it might have meant some
data would be incorrectly freed and objects would be left pointing into
dead memory.