forgot to mention this in the last commit
There was another small change made in the last commit that I forgot to
mention in the message, so I'll do so here. There was a conditional left
over that would cause us to not generate an empty globals array when
there were no global variables declared. This would cause the C compiler
to error as we were still using that array by passing it to
`gc_push_roots` at the start of the program. The previous commit removed
that.