initial support for treating operators as values
We create a new header file "operators.h" which is not used by the
runtime. Instead, the form `(operator +)` returns that operator as a
first-class value using the implementation `builtin_add` in the header
file.

Only the (+) operator is implemented at the moment; the rest will
produce compile errors for non-existent functions. Once the rest are
implemented, I will begin the work to make `(operator _)` and `(call _)`
unnecessary.