check for allocation failure while copying
At the moment, one might think it should be impossible for any attempt
to allocate on the new heap during garbage collection to fail, which
means we could skip checking it. After all, the point of garbage
collection is to remove unused objects and compact the rest - how could
this cause the new heap to end up even larger?

The answer is alignment. Imagine a heap that contains many living 1- and
2-byte objects which must be aligned to their respective sizes. Then
imagine that they are laid out like this:

        -------------------------------
        |a|b|c|d|e e|f f|g g|h h|. . .|
        -------------------------------

The GC sweeps these objects and copies them to the new heap in the order
that it visits them, which is potentially arbitrary. Thus the result
could be:

        -------------------------------
        |a .|e e|b .|f f|c .|g g|d .|h h
        -------------------------------

And so allocation will fail.

I don't think this error case was very likely to occur. It would only
happen if the GC was under high pressure and managed to visit objects in
a very unfortunate order. In fact, since I'm pretty sure every type of
object we're allocating on the heap right now has an alignment of 8
bytes on 64-bit systems, it might not be possible at all. However,
lacking access to a 32-bit system on which to check whether the
alignments of `Cons`, `HeapStr`, and `void*` are the same, I added this
check out of an abundance of caution.