add shared heap cell with (ref _) and (! _) fns
Of course, since there's no equivalent of `(:=)`, these operators are
currently not very useful. The main reason I added this is because I
intend to reuse the heap-allocated cell type to store captured variables
in closures, but it needs to be an object variant so it can be stored in
the local object array like everything else. The compiler will need to
specifically know to go through that layer of indirection when using
that variable in the function that originally defined it. Once closures
are implemented, I could remove these builtins, making these cell values
unobservable in normal user code.  I'm also considering using these cell
values to represent thunks.