add runtime support for closures
The closure data type now exists in the runtime! We still aren't letting
the user actually construct them with `(lambda)`, though - that will
have to wait.

Instead, in order to demonstrate that it works, I've added a `(counter)`
builtin that allocates and returns a new, very simple closure which
takes no arguments and returns successive integers every time it is
called. In a working Scheme implementation it would be something like
this:

    (define (counter)
      (define x 0)
      (lambda ()
        (define y x)
        (set! x (+ x 1))
        y))

While the implementation currently works, adding (counter) led me to
realize just how wasteful of memory it is. By itself, calling (counter)
allocates a staggering *56* bytes on the heap!

The reason for this is that every variable is stored in a separate cell,
and the closure struct itself just contains an array of pointers to
those cells. This means 2 unnecessary words of overhead per additional
variable in the closure's environment - one as part of the cell header,
and one for the pointer to that cell.

The original reason I had boxed variables be stored in their own cells
is because I imagined that different closures in a function scope might
capture different subsets of those variables. In that case, rather than
having them all try to share the same environment, they could only keep
pointers to the ones they needed, potentially allowing the GC to collect
the others.

Upon further reflection, though, that point doesn't really hold up. It
describes a very unusual case -- a function returning multiple
independent lambdas which capture different subsets of large variables
that the caller wants to drop at different times. The only place that
might happen is if the user was doing some kind of quasi-OO pattern, but
even in that case users don't generally expect some fields of an object
to be collected before others once they no longer call any methods that
access them.

The more sensible approach is to allocate all captured variables in a
single array, and then have every closure that needs any one of the
variables keep a pointer to that array. I was originally hesitant to do
this, partly because of the reasons described above, and partly because
I thought it would be harder to implement. I have a pretty good idea of
how to implement it now, though, since I've thought about it some more.

A few things to note about this approach:

- It might complicate the question of how to handle nested closures, by
  which I mean constructs like this:

    (let ((x 0))
      (lambda ()
        (lambda () x)))

  I think the inner closure would need to indicate that it captures the
  entire environment of its parent? Not entirely sure about that.

- Another downside of the current approach is that there are two levels
  of indirection between the `Obj` value representing the closure and
  its environment. The better approach doesn't solve that - I'm still
  going to be allocating a 3-word (one for the GC header, one for the
  function pointer, one for the environment pointer) value on the heap
  every time a closure is constructed.

  This could probably be fixed by storing it entirely inline, fitting
  one of the pointers into the object tag with some bit twiddling. I
  don't have the courage to try to do that now though - 3 words of
  overhead per closure is already much preferable to the 2-word-per-
  captured-variable status quo.