Rename `try_heap_alloc` to `retry_heap_alloc`
`try_` doesn't have the implication I wanted it to. In Rust at least,
the convention is to use that prefix to indicate that a function exits
early if it encounters a problem rather than blocking/panicking/etc. to
address it. `try_heap_alloc`'s relationship to the normal `heap_alloc`
is the opposite - it does try to address the problem, by performing a GC
cylce.