track local variables in single Map, not Map list
The whole advantage of immutable data structures is that you can update
them and cheaply keep the original. We take take advantage of this by
just overwriting the map bindings with the new variables introduced by a
`(let)`, rather than creating an entirely new map and consing it onto
the front.

The only tricky thing about this was that there was one place where we
matched on the list of scopes, checking if it was empty to determine if
the expression was being evaluated at the top level. This wasn't even a
correct way of determining that anyway, so I removed that check and
instead added a `top_level` parameter that we pass around explictly
(although it might be a good idea to group that with `block_level` into
some new record?)

In the process of doing this, I also fixed a bug that had to do with
expressions like this:

    (let ((x 1))
      (define x 2)
      x)

There are three ways this expression could plausibly be interpreted:

1. The two definitions of `x` refer to the same variable, which is
   assigned a new value by `(define)`. The result is 2.
2. The two definitions of `x` refer to different variables, and the one
   introduced by `(let)` takes precedence in the body. The result is 1.
3. The two definitions of `x` refer to different variables, and the one
   introduced by `(define)` is used in the body. The result is 2.

I'm pretty sure (1) goes against the standard. I haven't read it closely
enough to determine whether (2) or (3) is correct, but (3) definitely
seems more intuitive and is what the implementations I've tested do. Our
implementation formerly did (2) and now does (3).