add test harness and Makefile to call it
I've been using a Makefile set up in an untracked directory that would
compile and run an example file. So far this has been a good solution,
but now that there are quite a few complex features and I have to get
around to adding builtins, it will be useful to have a more mechanized
test suite.