only generate dynamic operator impls if used
Having to maintain a special `operators.h` header file was getting
tiring, especially since there are a large number of repetitve operators
(`car`, `cdar`, `cadar`, `cdadar`, etc) which all need to be separately
supported but which bloat compile times. Instead, we create an array of
flags for every operator, and set the flag only if it is used
dynamically (i.e. in non-function position).

We also reorder some of the builtins and their implementations to be
consistent with the order they appear in the spec, and remove the
builtin `(counter)` which isn't useful since it was only introduced to
test that closures worked.