add (set!)
This ended up being a pretty small change, requiring barely any
modification to existing code, in part because I'd been designing the
code generation of variables from the start with the understanding that
they would eventually be made mutable.

I didn't add any tests about the scoping rules of `(set!)`, since it
uses the exact same algorithm as variable lookup (in particular, we
don't allow using it to define new variables). The only difference is
that you cannot modify the built-in procedures - this is a weird way
in which builtins don't act exactly like user-defined procedures, but
the spec permits it in section 6:

> A program may use a top-level definition to bind any variable. It may
> subsequently alter such a binding by an assignment (see 4.1.6). These
> operations do not modify the behavior of Scheme's built-in procedures.
> Altering any top-level binding that has not been introduced by a
> definition has an unspecified effect on the behavior of the built-in
> procedures.