unify Str and HeapStr using `gc_tag` as const flag
Previously we had two structs, `Str` and `HeapStr`, which differed only
in that `HeapStr` contained a GC metadata field. Objects of type `Str`
held statically allocated string literals, while `HeapStr` was used for
strings created dynamically, although `(string-copy)` is currently the
only way to do this. The garbage collector relied on the difference
between these tags to avoid trying to treat static strings as garbage.

However, this approach complicated the type checking functionality and
would require I introduce static cons cell and vector types to deal with
objects created with `(quote)`. Rather than doing this, I decided to
mark objects as constant by setting their GC metadata pointer to a
sentinel value: `GC_SKIPME`.

`NULL` is already taken, so this value is currently `alignof(Obj) + 1`
in the hopes that this couldn't possibly be a valid pointer to an object
on one of the heaps. `1` could probably have worked equally well.